@@ -15,7 +15,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 2.0.0+2
 
 environment:
   sdk: '>=3.1.5 <4.0.0'
@@ -29,23 +29,23 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  get: ^4.6.1
-  device_frame: ^1.0.0
-  clipboard:
-  url_launcher: ^6.0.17
-  shimmer: ^2.0.0
-  charts_flutter: ^0.12.0
+  get: ^4.6.6
+  device_frame: ^1.1.0
+  url_launcher: ^6.2.6
   shared_preferences:
-  device_info_plus: ^3.2.1
-  flare_flutter: ^3.0.2
-  matrix4_transform: ^2.0.1
-  percent_indicator: ^3.4.0
+  device_info_plus: ^10.1.0
   intl:
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.8
+  responsive_framework: ^1.4.0
+  go_router: ^14.1.3
+  equatable: ^2.0.5
+  gap: ^3.0.1
+  flutter_staggered_grid_view: ^0.7.0
+  font_awesome_flutter: ^10.7.0
+  flutter_svg: ^2.0.10+1
 
-
+dependency_overrides:
+  file: ^7.0.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -71,9 +71,17 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
    - assets/
-   - assets/boltgrocery/
-   - assets/dailytodo/
-   - assets/legacyweather/
+   - assets/apps/crm/
+   - assets/apps/crm/screenshots/
+   - assets/apps/inventory/
+   - assets/apps/inventory/screenshots/
+   - assets/apps/fms/
+   - assets/apps/fms/screenshots/
+   - assets/apps/dis-aircon/
+   - assets/apps/dis-aircon/screenshots/
+   - assets/apps/try_modest/
+   - assets/apps/try_modest/screenshots/
+   - assets/stores/
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
 
