@@ -1,9 +1,65 @@
 import 'package:asadamatic/src/mvc/models/package.dart';
+import 'package:asadamatic/src/mvc/models/tool.dart';
 import 'package:asadamatic/src/mvc/models/value_type.dart';
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:get/get.dart';
 
 import 'package:flutter/foundation.dart';
+
+String stackOverReputationText = "Stackoverflow Reputation";
+String githubRepositoriesText = "Github Repositories";
+String commercialProjectsText = "Commercial Projects";
+
+const String playStoreBadge = "assets/stores/play-store.svg";
+const String appStoreBadge = "assets/stores/app-store.svg";
+
+const int stackoverflowScore = 2761;
+int get oneBySixtyOfScore => (stackoverflowScore / 60).round();
+
+const int githubRepoCount = 78;
+int get oneBySixOfCount => (stackoverflowScore / 600).round();
+
+const int commercialProjectsCount = 6;
+
+List<Score> scores = [
+  Score(
+      title: stackOverReputationText,
+      score: stackoverflowScore,
+      profileUrl: stackOverflowUrl),
+  Score(
+      title: githubRepositoriesText,
+      score: githubRepoCount,
+      profileUrl: githubUrl),
+  Score(title: commercialProjectsText, score: commercialProjectsCount),
+];
+const String githubUrl = 'https://github.com/asadamatic';
+const String stackOverflowUrl =
+    'https://stackoverflow.com/users/10285344/asad-hameed';
+
+const List<Tool> tools = [
+  Tool(name: 'Flutter', logoPath: 'assets/logos/flutter.png', slug: 'flutter'),
+  Tool(name: 'Dart', logoPath: 'assets/logos/dart.png', slug: 'dart'),
+  Tool(name: 'FastAPI', logoPath: 'assets/logos/fastapi.png', slug: 'fastapi'),
+  Tool(name: 'Python', logoPath: 'assets/logos/python.png', slug: 'python'),
+  Tool(name: 'Git', logoPath: 'assets/logos/git.png', slug: 'git'),
+  Tool(name: 'Github', logoPath: 'assets/logos/github.png', slug: 'github'),
+  Tool(name: 'GraphQL', logoPath: 'assets/logos/graphql.png', slug: 'graphql'),
+  Tool(name: 'SQLite', logoPath: 'assets/logos/sqlite.png', slug: 'sqlite'),
+  Tool(name: 'MySQL', logoPath: 'assets/logos/mysql.png', slug: 'mysql'),
+  Tool(
+      name: 'Firebase',
+      logoPath: 'assets/logos/firebase.png',
+      slug: 'firebase'),
+  Tool(
+      name: 'Android Studio',
+      logoPath: 'assets/logos/android_studio.png',
+      slug: 'android-studio'),
+  Tool(name: 'VS Code', logoPath: 'assets/logos/vscode.png', slug: 'vs-code'),
+  Tool(
+      name: 'Selenium',
+      logoPath: 'assets/logos/selenium.png',
+      slug: 'selenium'),
+];
 
 class AppConstants {
   static const String appTitle = 'Asad Hameed - Flutter Developer';
@@ -12,54 +68,6 @@ class AppConstants {
       r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+");
 
   // Scores
-
-  static const List<List<dynamic>> scores = [
-    ["Stackoverflow Reputation", ValueType.stackoverflow],
-    ["Github Repositories", ValueType.github],
-    ["Commercial Projects", ValueType.commercialProjects]
-  ];
-
-  static const int stackoverflowScore = 2416;
-  static int get oneBySixtyOfScore => (stackoverflowScore / 60).round();
-
-  static const int githubRepoCount = 78;
-  static int get oneBySixOfCount => (stackoverflowScore / 600).round();
-
-  static const int commercialProjects = 6;
-
-  static const List<List<String>> descriptions = [
-    [
-      "Grocery",
-      "An app that lets You create grocery lists, so You never forget an item! "
-    ],
-    ['Todo', 'Productivity Simplified with daily todo app!'],
-    [
-      'Weather',
-      'Stay up to date with weather conditions with Weather by Legacy'
-    ]
-  ];
-  static const List<String> os = ['Android', 'Ios'];
-  static const List<List<String>> socialIcons = [
-    ['assets/github.png', 'https://github.com/asadamatic'],
-    [
-      'assets/stack-overflow.png',
-      'https://stackoverflow.com/users/10285344/asad-hameed'
-    ],
-    ['assets/linkedin.png', 'https://www.linkedin.com/in/asadamatic']
-  ];
-  static const String linkedInUrl =
-      'https://stackoverflow.com/users/10285344/asad-hameed';
-  static const List<String> appLogoIcons = [
-    'assets/boltgrocery/icon.png',
-    'assets/dailytodo/icon.png',
-    'assets/legacyweather/icon.png'
-  ];
-
-  static const List<String> packages = [
-    'firestore_search',
-    'file_case',
-    'dialogs'
-  ];
 
   static const List<Package> packagesDescription = [
     Package(
@@ -83,19 +91,6 @@ class AppConstants {
         pubPoints: '110',
         popularity: '80',
         likes: '12'),
-  ];
-  static const List<String> tools = [
-    'Flutter',
-    'FastAPI',
-    'Python',
-    'Java',
-    'JavaFx',
-    'SQLite',
-    'MySQL',
-    'Firebase',
-    'Android',
-    'C++',
-    'Selenium',
   ];
 
   static DeviceInfoPlugin get deviceInfo => DeviceInfoPlugin();
@@ -135,3 +130,115 @@ class AppConstants {
   //   return desktopType;
   // }
 }
+
+const String fmsLogo = 'assets/apps/fms/logo.png';
+const String fmsScreenshotSplash = 'assets/apps/fms/screenshots/splash.png';
+const String fmsScreenshotCheckinout =
+    'assets/apps/fms/screenshots/checkinout.png';
+const String fmsScreenshotHome = 'assets/apps/fms/screenshots/home.png';
+const String fmsScreenshotNotifications =
+    'assets/apps/fms/screenshots/notifications.png';
+const String fmsScreenshotProduct = 'assets/apps/fms/screenshots/product.png';
+const String fmsScreenshotSchedule = 'assets/apps/fms/screenshots/schedule.png';
+const String fmsScreenshotTasks = 'assets/apps/fms/screenshots/tasks.png';
+const String fmsScreenshotWorkOrder =
+    'assets/apps/fms/screenshots/work_order.png';
+const String fmsScreenshotTaskDetail =
+    'assets/apps/fms/screenshots/task_detail.png';
+
+const List<String> fmsScreenshots = [
+  fmsScreenshotSplash,
+  fmsScreenshotHome,
+  fmsScreenshotTasks,
+  fmsScreenshotSchedule,
+  fmsScreenshotTaskDetail,
+  fmsScreenshotProduct,
+  fmsScreenshotWorkOrder,
+  fmsScreenshotCheckinout,
+  fmsScreenshotNotifications,
+];
+const String inventoryInternalTransfers =
+    'assets/apps/inventory/screenshots/internal_transfers.png';
+const String inventoryFilters = 'assets/apps/inventory/screenshots/filters.png';
+const String inventoryPurchaseOrders =
+    'assets/apps/inventory/screenshots/purchase_orders.png';
+const String inventoryPurchaseOrder =
+    'assets/apps/inventory/screenshots/purchase_order.png';
+const String inventoryHome = 'assets/apps/inventory/screenshots/home.png';
+const String inventoryScanner = 'assets/apps/inventory/screenshots/scanner.png';
+const String inventorySplash = 'assets/apps/inventory/screenshots/splash.png';
+
+const List<String> inventoryScreenshots = [
+  inventorySplash,
+  inventoryHome,
+  inventoryInternalTransfers,
+  inventoryFilters,
+  inventoryPurchaseOrders,
+  inventoryPurchaseOrder,
+  inventoryScanner,
+];
+
+const String disAirconAppointments =
+    'assets/apps/dis-aircon/screenshots/appointments.png';
+const String disAirconHome = 'assets/apps/dis-aircon/screenshots/home.png';
+const String disAirconOrder = 'assets/apps/dis-aircon/screenshots/order.png';
+const String disAirconProducts =
+    'assets/apps/dis-aircon/screenshots/products.png';
+const String disAirconServicingContracts =
+    'assets/apps/dis-aircon/screenshots/servicing_contracts.png';
+const String disAirconSplash = 'assets/apps/dis-aircon/screenshots/splash.png';
+
+const List<String> disAirconScreenshots = [
+  disAirconSplash,
+  disAirconHome,
+  disAirconProducts,
+  disAirconServicingContracts,
+  disAirconAppointments,
+  disAirconOrder,
+];
+
+const String crmSplash = 'assets/apps/crm/screenshots/splash.png';
+const String crmAddCall = 'assets/apps/crm/screenshots/add_call.png';
+const String crmAddMeeting = 'assets/apps/crm/screenshots/add_meeting.png';
+const String crmAudioNote = 'assets/apps/crm/screenshots/audio_note.png';
+const String crmCallScreen = 'assets/apps/crm/screenshots/call_screen.png';
+const String crmDialpad = 'assets/apps/crm/screenshots/dialpad.png';
+const String crmHome = 'assets/apps/crm/screenshots/home.png';
+const String crmLeadProfile = 'assets/apps/crm/screenshots/lead_profile.png';
+const String crmLeads = 'assets/apps/crm/screenshots/leads.png';
+
+const List<String> crmScreenshots = [
+  crmSplash,
+  crmHome,
+  crmLeads,
+  crmLeadProfile,
+  crmAudioNote,
+  crmDialpad,
+  crmCallScreen,
+  crmAddCall,
+  crmAddMeeting,
+];
+
+const String tryModestProduct =
+    'assets/apps/try_modest/screenshots/product.png';
+const String tryModestProducts =
+    'assets/apps/try_modest/screenshots/products.png';
+const String tryModestHome = 'assets/apps/try_modest/screenshots/home.png';
+const String tryModestCollections =
+    'assets/apps/try_modest/screenshots/collections.png';
+const String tryModestSplash = 'assets/apps/try_modest/screenshots/splash.png';
+
+const String tryModestAddresses =
+    'assets/apps/try_modest/screenshots/addresses.png';
+const String tryModestProfile =
+    'assets/apps/try_modest/screenshots/profile.png';
+
+const List<String> tryModestScreenshots = [
+  tryModestSplash,
+  tryModestHome,
+  tryModestCollections,
+  tryModestProducts,
+  tryModestProduct,
+  tryModestProfile,
+  tryModestAddresses
+];
