@@ -1,15 +1,20 @@
 import 'package:asadamatic/src/mvc/controllers/home_controller.dart';
 import 'package:asadamatic/src/mvc/models/package.dart';
-import 'package:asadamatic/src/style/values.dart';
+import 'package:asadamatic/src/style/styles.dart';
+import 'package:asadamatic/src/widgets/responsive_grid_view_builder.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:shimmer/shimmer.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 class PackagesSection extends StatelessWidget {
   const PackagesSection({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final textTheme = theme.textTheme;
+    final colorScheme = theme.colorScheme;
+
     return GetBuilder<HomeController>(
         id: 'updatePackagesData',
         builder: (_controller) {
@@ -17,73 +22,21 @@ class PackagesSection extends StatelessWidget {
               (BuildContext layoutBuilderContext, BoxConstraints constraints) {
             return Column(
               children: [
-                const Flexible(
-                    child: SizedBox(
-                  height: 100.0,
-                )),
                 Text(
                   'Contributions to Flutter',
                   textAlign: TextAlign.center,
-                  style: Theme.of(context).textTheme.headlineMedium,
+                  style: textTheme.headlineMedium,
                 ),
-                const SizedBox(
-                  height: 30.0,
-                ),
-                constraints.maxWidth < 780
-                    ? Column(
-                        crossAxisAlignment: CrossAxisAlignment.center,
-                        children: _controller.packagesDataLoaded
-                            ? _controller.packagesData
-                                .map((packageData) => PackageCard(
-                                      package: packageData,
-                                    ))
-                                .toList()
-                            : const [
-                                PackageShimmer(),
-                                PackageShimmer(),
-                                PackageShimmer()
-                              ])
-                    : constraints.maxWidth < 1170
-                        ? Column(
-                            children: [
-                              Row(
-                                  mainAxisAlignment: MainAxisAlignment.center,
-                                  children: _controller.packagesDataLoaded
-                                      ? [
-                                          PackageCard(
-                                            package:
-                                                _controller.packagesData[0],
-                                          ),
-                                          PackageCard(
-                                            package:
-                                                _controller.packagesData[1],
-                                          ),
-                                        ]
-                                      : const [
-                                          PackageShimmer(),
-                                          PackageShimmer()
-                                        ]),
-                              _controller.packagesDataLoaded
-                                  ? PackageCard(
-                                      package: _controller.packagesData[2],
-                                    )
-                                  : const PackageShimmer()
-                            ],
-                          )
-                        : Row(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: _controller.packagesDataLoaded
-                                ? _controller.packagesData
-                                    .map((packageData) => PackageCard(
-                                          package: packageData,
-                                        ))
-                                    .toList()
-                                : const [
-                                    PackageShimmer(),
-                                    PackageShimmer(),
-                                    PackageShimmer()
-                                  ],
-                          ),
+                CustomResponsiveGridView(
+                    phoneCrossAxisCount: 1,
+                    tabletCrossAxisCount: 2,
+                    padding: EdgeInsets.symmetric(horizontal: massiveSpacing),
+                    physics: const NeverScrollableScrollPhysics(),
+                    shrinkWrap: true,
+                    itemBuilder: (context, index) => PackageCard(
+                          package: _controller.packagesData[index],
+                        ),
+                    itemCount: _controller.packagesData.length)
               ],
             );
           });
@@ -92,206 +45,81 @@ class PackagesSection extends StatelessWidget {
 }
 
 class PackageCard extends StatelessWidget {
-  const PackageCard({Key? key, this.package}) : super(key: key);
-  final Package? package;
+  const PackageCard({Key? key, required this.package}) : super(key: key);
+  final Package package;
   @override
   Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final textTheme = theme.textTheme;
+    final colorScheme = theme.colorScheme;
+
     return Card(
-      elevation: 2.0,
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8.0)),
+      elevation: smallElevation,
+      shape: RoundedRectangleBorder(borderRadius: borderRadius),
       margin: const EdgeInsets.all(15.0),
-      child: Container(
-          width: 300.0,
-          height: 250.0,
-          padding: const EdgeInsets.all(28.0),
+      child: Padding(
+          padding: const EdgeInsets.all(hugeSpacing),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             crossAxisAlignment: CrossAxisAlignment.start,
             mainAxisSize: MainAxisSize.min,
             children: [
-              Text(
-                package!.name!,
-                style: TextStyle(
-                    fontSize: Theme.of(context).textTheme.titleLarge!.fontSize),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Text(
+                    package.name!,
+                    style: TextStyle(fontSize: textTheme.titleLarge!.fontSize),
+                  ),
+                  IconButton(
+                      onPressed: () => launchUrl(Uri.parse(package.url)),
+                      icon: const Icon(Icons.open_in_new_rounded))
+                ],
               ),
               const SizedBox(
-                height: 5.0,
+                height: smallSpacing,
               ),
-              SizedBox(
-                width: 270.0,
-                height: 60.0,
-                child: Text(
-                  package!.description!,
+              Text(package.description!,
                   overflow: TextOverflow.ellipsis,
                   maxLines: 3,
-                  style: TextStyle(
-                      height: 1.5,
-                      fontSize:
-                          Theme.of(context).textTheme.bodyLarge!.fontSize),
-                ),
-              ),
+                  style: textTheme.bodyLarge),
               const Flexible(child: SizedBox(height: 30.0)),
-              Row(
+              Table(
+                columnWidths: const <int, TableColumnWidth>{
+                  0: IntrinsicColumnWidth(),
+                  1: IntrinsicColumnWidth(),
+                  2: IntrinsicColumnWidth(),
+                },
                 children: [
-                  const Flexible(
-                    child: SizedBox(
-                      width: 50.0,
-                    ),
-                  ),
-                  Table(
-                    columnWidths: const <int, TableColumnWidth>{
-                      0: IntrinsicColumnWidth(),
-                      1: IntrinsicColumnWidth(),
-                      2: IntrinsicColumnWidth(),
-                    },
-                    children: [
-                      TableRow(
-                          children: [
-                        package!.likes!,
-                        package!.pubPoints!,
-                        package!.popularity!
-                      ]
-                              .map(
-                                (value) => Padding(
-                                  padding: const EdgeInsets.only(
-                                      top: 8.0, right: 8.0, left: 8.0),
-                                  child: Text(value,
-                                      style: Theme.of(context)
-                                          .textTheme
-                                          .bodyMedium),
-                                ),
-                              )
-                              .toList()),
-                      TableRow(
-                          children: ['Likes', 'Pub Points', 'Popularity']
-                              .map((label) => Padding(
-                                    padding: const EdgeInsets.all(8.0),
-                                    child: Text(label,
-                                        style: Theme.of(context)
-                                            .textTheme
-                                            .bodyMedium),
-                                  ))
-                              .toList())
-                    ],
-                  ),
+                  TableRow(
+                      children: [
+                    package.likes,
+                    package.pubPoints,
+                    package.popularity!
+                  ]
+                          .map(
+                            (value) => Padding(
+                              padding: const EdgeInsets.only(
+                                  top: 8.0, right: 8.0, left: 8.0),
+                              child: Text(value!,
+                                  style:
+                                      Theme.of(context).textTheme.bodyMedium),
+                            ),
+                          )
+                          .toList()),
+                  TableRow(
+                      children: ['Likes', 'Pub Points', 'Popularity']
+                          .map((label) => Padding(
+                                padding: const EdgeInsets.all(8.0),
+                                child: Text(label,
+                                    style:
+                                        Theme.of(context).textTheme.bodyMedium),
+                              ))
+                          .toList())
                 ],
               )
             ],
           )),
     );
-  }
-}
-
-class PackageShimmer extends StatelessWidget {
-  const PackageShimmer({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-        elevation: 2.0,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8.0)),
-        margin: const EdgeInsets.all(15.0),
-        child: Container(
-            width: 300.0,
-            height: 250.0,
-            padding: const EdgeInsets.all(28.0),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Shimmer.fromColors(
-                  child: Container(
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(4.0),
-                      color: Colors.grey[200]!,
-                    ),
-                    height: 20.0,
-                    width: 160.0,
-                  ),
-                  baseColor: AppStyles.shimmerBaseColor,
-                  highlightColor: AppStyles.shimmerHighlightColor,
-                ),
-                const SizedBox(
-                  height: 5.0,
-                ),
-                SizedBox(
-                    width: 270.0,
-                    height: 60.0,
-                    child: Column(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [1, 2, 3, 4, 5]
-                            .map((index) => index % 2 == 0
-                                ? const SizedBox(
-                                    height: 6.0,
-                                  )
-                                : Shimmer.fromColors(
-                                    child: Container(
-                                      decoration: BoxDecoration(
-                                        borderRadius:
-                                            BorderRadius.circular(4.0),
-                                        color: Colors.grey[200]!,
-                                      ),
-                                      height: 14.0,
-                                      width: 270.0,
-                                    ),
-                                    baseColor: AppStyles.shimmerBaseColor,
-                                    highlightColor:
-                                        AppStyles.shimmerHighlightColor,
-                                  ))
-                            .toList())),
-                const Flexible(child: SizedBox(height: 30.0)),
-                Row(
-                  children: [
-                    const Flexible(
-                      child: SizedBox(
-                        width: 50.0,
-                      ),
-                    ),
-                    Table(
-                      columnWidths: const <int, TableColumnWidth>{
-                        0: IntrinsicColumnWidth(),
-                        1: IntrinsicColumnWidth(),
-                        2: IntrinsicColumnWidth(),
-                      },
-                      children: [
-                        TableRow(
-                            children: [1, 2, 3]
-                                .map((index) => Padding(
-                                      padding: const EdgeInsets.only(
-                                          top: 8.0, right: 8.0, left: 8.0),
-                                      child: Shimmer.fromColors(
-                                        child: Container(
-                                          height: 14.0,
-                                          width: 20.0,
-                                          decoration: BoxDecoration(
-                                            borderRadius:
-                                                BorderRadius.circular(4.0),
-                                            color: Colors.grey[200]!,
-                                          ),
-                                        ),
-                                        baseColor: AppStyles.shimmerBaseColor,
-                                        highlightColor:
-                                            AppStyles.shimmerHighlightColor,
-                                      ),
-                                    ))
-                                .toList()),
-                        TableRow(
-                            children: ['Likes', 'Pub Points', 'Popularity']
-                                .map((label) => Padding(
-                                      padding: const EdgeInsets.all(8.0),
-                                      child: Text(label,
-                                          style: Theme.of(context)
-                                              .textTheme
-                                              .bodyMedium),
-                                    ))
-                                .toList())
-                      ],
-                    ),
-                  ],
-                )
-              ],
-            )));
   }
 }
