@@ -1,7 +1,10 @@
-import 'package:asadamatic/src/constant/values.dart';
+import 'package:asadamatic/src/mvc/controllers/home_controller.dart';
+import 'package:asadamatic/src/mvc/models/app_data.dart';
+import 'package:asadamatic/src/mvc/models/social_link.dart';
 import 'package:asadamatic/src/widgets/app_logo_icon.dart';
 import 'package:asadamatic/src/widgets/social_logo_icon.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class SocialPalette extends StatelessWidget {
   const SocialPalette({Key? key}) : super(key: key);
@@ -10,11 +13,9 @@ class SocialPalette extends StatelessWidget {
   Widget build(BuildContext context) {
     return Column(
       mainAxisAlignment: MainAxisAlignment.center,
-      children: AppConstants.socialIcons
-          .map((socialIcon) => SocialLogoIcon(
-                asset: socialIcon[0],
-                index: AppConstants.socialIcons.indexOf(socialIcon),
-                url: socialIcon[1],
+      children: socialLinks
+          .map((link) => SocialLogoIcon(
+                socialLink: link,
               ))
           .toList(),
     );
@@ -26,15 +27,38 @@ class AppIconPalette extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: AppConstants.appLogoIcons
-          .map((appLogoIcon) => AppLogoIcon(
-                asset: appLogoIcon,
-                index: AppConstants.appLogoIcons.indexOf(appLogoIcon),
-                app: appLogoIcon.split('/')[1],
-              ))
-          .toList(),
-    );
+    return GetBuilder<HomeController>(
+        id: 'apps',
+        builder: (controller) {
+          return Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              IconButton(
+                  onPressed: controller.moveAppsBackwards,
+                  icon: const Icon(Icons.arrow_back_ios_new_outlined)),
+              SizedBox(
+                height: 100.0,
+                width: 100.0,
+                child: PageView.builder(
+                    controller: controller.appsPageController,
+                    scrollDirection: Axis.horizontal,
+                    // itemCount: apps.length,
+                    onPageChanged: (index) =>
+                        controller.onChangedApp(index % apps.length),
+                    itemBuilder: (context, index) {
+                      final adjustedIndex = index % apps.length;
+                      final app = apps[adjustedIndex];
+                      return AppLogoIcon(
+                        isSelected: controller.selectedApp.value == app,
+                        app: app,
+                      );
+                    }),
+              ),
+              IconButton(
+                  onPressed: controller.moveAppsForwards,
+                  icon: const Icon(Icons.arrow_forward_ios_outlined)),
+            ],
+          );
+        });
   }
 }
